+++ scripts/downscale_bc.py
@@ -21,17 +21,17 @@ period_future = snakemake.params.period_future
 period_hist = snakemake.params.period_hist
 
 # test
-# proj_file="results/projections/_CORDEX_AFR-22_CLMcom-KIT_NCC-NorESM1-M_rcp85_r1i1p1_CCLM5-0-15_v1_chelsa2_done.txt"
+# proj_file="results/projections/_CMIP6_world_AWI_AWI-CM-1-1-MR_ssp126_r1i1p1f1_none_none_chelsa2_done.txt"
 # base_hist_file="results/baselines/Côte-d'Ivoire_chelsa2_monthly-means_1980-2005.nc"
 # area="Côte-d'Ivoire"
-# project="CORDEX"
-# domain="AFR-22"
-# institute="CLMcom-KIT"
-# model="NorESM1-M"
-# experiment="rcp85"
-# ensemble="r1i1p1"
-# rcm="CCLM5-0-15"
-# downscaling="v1"
+# project="CMIP6"
+# domain="world"
+# institute="AWI"
+# model="AWI-CM-1-1-MR"
+# experiment="ssp126"
+# ensemble="r1i1p1f1"
+# rcm="none"
+# downscaling="none"
 # baseline="chelsa2"
 # aggregation="monthly-means"
 # period_future="2071-2100"
@@ -52,7 +52,7 @@ base_hist = xr.open_mfdataset(base_hist_file, parallel=True)
 # anomalies
 anomalies = proj_future - proj_hist
 if 'pr' in list(proj_future.keys()):
-  anomalies_rel =  (proj_future - proj_hist)/proj_hist
+  anomalies_rel =  (proj_future - proj_hist)/(proj_hist+1)
   anomalies["pr"] = anomalies_rel["pr"]
 
 # add to the baseline